Fix field-renderer header and label in schema mode
