+++ packages/cardhost/app/components/right-edge.js
@@ -1,5 +1,5 @@
 import Component from '@glimmer/component';
-import { fieldComponents } from './card-manipulator';
+import { fieldComponents } from '@cardstack/core/utils/mappings';
 import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
