+++ packages/core/app/utils/mappings.js
@@ -1,0 +1,1 @@
+export * from '@cardstack/core/utils/mappings';
